// yiqiheshebei/src/service/instrument-manage/lendRegistration/types.ts
export interface RequestDataLabName {
  data: {
    id: number
    title: string
  }[]
}

// 实验室批量配置保存
export interface StoreItemType {
  id: number
  number_plies: number
  is_del: number
  numbers: string
  qr_code: string
  school_id: number
  school_laboratory_id: number
  school_laboratory_name: string
  subject_id: null
  title: string
  type: number
  user_id: number
  user_name: string
  url: string
}

export const RequestDataLab: RequestDataLabName = {
  data: [
    {
      id: 1,
      title: ''
    }
  ]
}

// 出借列表
export interface lendList {
  total: number
  data: {
    id: number
    record_number: string //出借单号
    school_id: number
    lender: string //出借人
    lend_id: number
    lend_desc: string
    start_time: string
    plan_end_time: string
    user_name: string
    status: number
    create_time: string
    is_del: number
    end_time: string
    total: number
    species: number
    list: {
      id?: number
      status: number
      name: string
      is_del: number
      specification: string
      return_total: number //已归还仪器数量
      total: number //出借总数
    }[]
  }[]
}
// ...
export interface addLendInstrumnetInfoList {
  lender: string
  lender_id: number | string
  type: number
  lend_desc: string
  start_time: string
  plan_end_time: string
  list: {
    hardware_number: string
    asset_number: string
    name: string
    bacth: string
    specification: string
    store_name: string
    instrument_id: number
    laboratory_id: number
    store_id: number
    number_plies: number
    total: number
  }[]
}
// ...
export interface getSchoolTeacherList {
  username: string
  id: number
  role: number
}

export const schoolTeacherList: getSchoolTeacherList = {
  username: '',
  id: 1,
  role: 1
}

// 出借仪器（仪器信息查询）
export interface instrumnetList {
  id: number
  asset_number: string
  bacth: string
  specification: string
  number: string
  name: string
  stockStatus: boolean
  weekList: []
  picture: string
  detail: {
    id: number
    instrumnet_id: number
    laboratory_name: string
    name: string
    number_plies: number
    store_id: number
    store_name: string
    total: number
  }[]
}
// ...
export interface lendDetailList {
  id: number
  record_number: string
  type: number
  lend_desc: string
  lender: string
  plan_end_time: string
  start_time: string
  end_time: string
  return_tips: string
  status: number
  is_del: number
  is_finish: number
  nuReturn: {
    id: number
    asset_number: string
    name: string
    residue: number
    damage_total: number
    scrap_total: number
    maintain_total: number
    total: number
    laboratory_name: string
    store_name: string
    number_plies: number
  }[]
  returnList: {
    id: number
    asset_number: string
    name: string
    total: number
    laboratory_name: string
    store_name: string
    number_plies: string | number
  }[]
  list: {
    id: number
    name: string
    asset_number: string
    total: number
    laboratory_name: string
    store_name: string
    number_plies: number
  }[]
}
// ...
export interface lendDetailInventoryList {
  create_time: string
  list: {
    name: string
    id: number
    total: number
    asset_number: string
    bacth: string
  }[]
}

export const lendDetailInventoryData: lendDetailInventoryList = {
  create_time: '',
  list: [
    {
      id: 1,
      name: '',
      asset_number: '',
      total: 0,
      bacth: ''
    }
  ]
}

// 仪器归还登记详情
export interface returnRegistrationList {
  id: number
  is_del: number
  is_finish: number
  lender: string
  end_time: string
  record_number: string
  plan_end_time: string
  start_time: string
  return_tips: string
  lend_desc: string
  status: number
  list: {
    id: number
    asset_number: string
    name: string
    bacth: string
    laboratory_id: number
    laboratory_name: string
    school_instrument_id: number
    store_id: number
    store_name: string
    number_plies: number
    total: number
    returnStatus: number
  }[]
  nuReturn: {
    id: number
    asset_number: string
    name: string
    bacth: string
    laboratory_id: number
    laboratory_name: string
    school_instrument_id: number
    store_id: number
    unReturnNum: number
    store_name: string
    number_plies: number
    total: number
    returnStatus: number
  }[]
  returnList: {
    id: number
    asset_number: string
    name: string
    bacth: string
    laboratory_id: number
    laboratory_name: string
    store_id: number
    store_name: string
    number_plies: number
    total: number
    returnStatus: number
    school_instrument_id: number
    detail: {
      // id: number
      // asset_number: string
      // name: string
      id: number
      Pid: number
      bacth: string
      laboratory_id: number
      laboratory_name: string
      store_id: number
      store_name: string
      number_plies: number
      total: number
    }[]
  }[]
}
// ...
export interface returnInfoList {
  id: number //记录单
  status: number
  return_tips: string
  list: {
    id: number
    school_instrument_id: number
    school_laboratory_id: number
    store_id: number
    number_plies: number
    bacth: string
    total: number //借用总数
    return_total: number //归还总数
  }[]
}

export const returnInfoData: returnInfoList = {
  id: 1,
  status: 1,
  return_tips: '',
  list: [
    {
      id: 1,
      school_instrument_id: 1,
      school_laboratory_id: 1,
      store_id: 1,
      number_plies: 1,
      bacth: '',
      total: 0,
      return_total: 0
    }
  ]
}

// 仪器归还转维修保养登记
export interface repairRegistrationList {
  status: number
  type: number
  mode: number
  factory: string
  start_time: string
  plan_end_time: string
  desc: string
  source_record_number: string
  source_type: number
  url: Array<string>
  list: {
    id: number
    bacth: string
    total: number
    loanTotal: number
    asset_number: string
    name: string
    laboratory_name: string
    store_name: string
    number_plies: string | number
  }[]
}
// ...
export interface scrapRegistrationList {
  status: number
  type: number
  mode: number
  cause: number
  process_mode: number
  desc: string
  source_type: number
  price: number | string
  source_record_number: string
  url: Array<string>
  list: {
    id: number
    bacth: string
    total: number
    loanTotal: number
    asset_number: string
    name: string
    laboratory_name: string
    store_name: string
    number_plies: string | number
  }[]
}
// ...
export interface damageRegistrationList {
  status: number
  type: number
  repay_man: number | string
  repay_id: number | string
  repay_time: string
  repay_desc: string
  damage_man: number | string
  damage_id: number | string
  repay_price: number | string
  desc: string
  source_type: number
  source_record_number: string
  url: Array<string>
  list: {
    id: number
    store_id: number
    laboratory_id: number
    number_plies: number | string
    asset_number: string
    name: string
    bacth: string
    total: number
    loanTotal: number
    laboratory_name: string
    store_name: string
  }[]
}